Update cut_plotter tests, add validators to LineEdits
Split cut_presenter_test into cut_plotter_presenter_test and
cut_widget_presenter test to match new class structure. Also
removed redundant tests (e.g. validation tested in axis_test).

Also added QValidators to slice and cut QLineEdits so that
non-numeric values cannot be entered. This means this case
no longer has to be tested.
