+++ mslice/widgets/cut/cut.py
@@ -5,11 +5,13 @@
 # -----------------------------------------------------------------------------
 from __future__ import (absolute_import, division, print_function)
 
+from mslice.util.qt import load_ui
+from mslice.util.qt.QtGui import QDoubleValidator
 from mslice.util.qt.QtCore import Signal
 from mslice.util.qt.QtWidgets import QWidget
 
 from mslice.presenters.cut_widget_presenter import CutWidgetPresenter
-from mslice.util.qt import load_ui
+
 from mslice.views.interfaces.cut_view import CutView
 from .command import Command
 
@@ -37,6 +39,7 @@ class CutWidget(CutView, QWidget):
         self._minimumStep = None
         self.lneCutStep.editingFinished.connect(self._step_edited)
         self.enable_integration_axis(False)
+        self.set_validators()
 
     def _btn_clicked(self):
         sender = self.sender()
@@ -49,10 +52,11 @@ class CutWidget(CutView, QWidget):
         if self._minimumStep:
             try:
                 value = float(self.lneCutStep.text())
+                print(value)
             except ValueError:
-                value = 0
+                value = 0.0
                 self.display_error('Invalid cut step parameter. Using default.')
-            if value == 0:
+            if value == 0.0:
                 self.lneCutStep.setText('%.5f' % (self._minimumStep))
                 self.display_error('Setting step size to default.')
             elif value < (self._minimumStep / 100.):
@@ -250,6 +254,13 @@ class CutWidget(CutView, QWidget):
 
         self.btnCutPlot.setEnabled(True)
         self.btnCutPlotOver.setEnabled(True)
+
+    def set_validators(self):
+        line_edits = [self.lneCutStart, self.lneCutEnd, self.lneCutIntegrationStart,
+                      self.lneCutIntegrationEnd, self.lneCutIntegrationWidth, self.lneEditCutIntensityStart,
+                      self.lneCutIntensityEnd]
+        for line_edit in line_edits:
+            line_edit.setValidator(QDoubleValidator())
 
     def force_normalization(self):
         self.rdoCutNormToOne.setEnabled(False)
